requirements.txt
discord.py==2.3.2
pillow==10.3.0